removing empty figures
