--- lineage/figures/figure10.py
@@ -1,16 +1,0 @@
-"""
-File: figure10.py
-Purpose: Generates figure 10.
-"""
-
-from .figureCommon import getSetup
-
-
-def makeFigure():
-    """
-    Makes figure 10.
-    """
-
-    _, f = getSetup((21, 6), (1, 3))
-
-    return f
